requirements.txt
requests
pandas
numpy
scikit-learn
tqdm
progressbar
matplotlib
pillow
